commit  by falak mvn

+++ src/test/java/org/apache/commons/codec/language/bm/RuleTest.java
@@ -14,6 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.commons.codec.language.bm;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -48,7 +49,8 @@ public class RuleTest {
             for (int i = 0; i < phs.length; i++) {
                 for (int j = i + 1; j < phs.length; j++) {
                     final int c = Rule.Phoneme.COMPARATOR.compare(phs[i], phs[j]);
-                    assertTrue(c < 0,
+
+                    assertTrue(Integer.valueOf(c).intValue() < 0,
                             "Comparing " + phs[i].getPhonemeText() + " to " + phs[j].getPhonemeText() + " should be negative");
                 }
             }
